seventh commit

+++ week_1/hirepurchase.py
@@ -4,8 +4,8 @@
 
 
 dep = float(input("Enter the deposited amount : "))
-n = float(input("Enter the number of months : "))
-inst = float(input("Enter the monthly installment paid : "))
+n = int(input("Enter the number of months : "))
+inst = int(input("Enter the monthly installment paid : "))
 
 hpp = dep + (n * inst)
